整合sql文件，增加累积月份表,增加dataset表字段
