--- alembic/versions/864ccb4da84a_增加数据总量表.py
@@ -1,38 +1,0 @@
-"""增加数据总量表
-
-Revision ID: 864ccb4da84a
-Revises: ee8f1867aa17
-Create Date: 2024-05-23 10:02:13.231810
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '864ccb4da84a'
-down_revision = 'ee8f1867aa17'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('cumulative_data_per_month',
-    sa.Column('date', sa.Date(), nullable=False, comment='日期'),
-    sa.Column('full_data_size', sa.Float(), nullable=True, comment='数据总量(GB)'),
-    sa.Column('full_data_count', sa.Float(), nullable=True, comment='数据条目'),
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False, comment='主键'),
-    sa.Column('gmt_create', sa.DateTime(), server_default=sa.text('now()'), nullable=False, comment='创建时间'),
-    sa.Column('gmt_modified', sa.DateTime(), server_default=sa.text('now()'), nullable=False, comment='修改时间'),
-    sa.Column('is_deleted', sa.Boolean(), server_default=sa.text('false'), nullable=False, comment='该行是否被删除'),
-    sa.PrimaryKeyConstraint('id'),
-    comment='数据总量'
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('cumulative_data_per_month')
-    # ### end Alembic commands ###
