--- alembic/versions/8d47b0117044_modified_dataset.py
@@ -1,30 +1,0 @@
-"""modified_dataset
-
-Revision ID: 8d47b0117044
-Revises: 864ccb4da84a
-Create Date: 2024-05-30 15:37:35.724576
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '8d47b0117044'
-down_revision = '864ccb4da84a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('dataset', sa.Column('planed_download_per_month', sa.Float(), nullable=True, comment='每月计划下载量'))
-    op.add_column('dataset', sa.Column('is_cleaned', sa.Boolean(), nullable=True, comment='是否清洗过数据'))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('dataset', 'is_cleaned')
-    op.drop_column('dataset', 'planed_download_per_month')
-    # ### end Alembic commands ###
